adiciona circular progress e otimizei os imports
