+++ semana12/labeddit/src/screens/PostPage/PostPage.js
@@ -1,9 +1,8 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { useParams } from 'react-router-dom';
 import { useRequestData } from '../../hooks/useRequestData';
-import { BASE_URL, HEADERS } from '../../constants/requestConfig';
+import { BASE_URL } from '../../constants/requestConfig';
 import PostCard from '../../components/PostComponents/PostCard/PostCard';
-import PostCardMaterial from '../../components/PostComponents/PostCard/PostCardMaterial';
 import { FlexBox } from '../../global/global-styles';
 import CircularProgress from '@material-ui/core/CircularProgress';
 
@@ -18,8 +17,9 @@ const PostPage = () => {
 
     return (
         <FlexBox>
+            {!data && <CircularProgress color="primary"/>}
             {data && 
-                <PostCardMaterial
+                <PostCard
                     upDate={getData}
                     key={data.post.id}
                     post={data.post}
